requirements.txt
captcha
tensorflow
numpy
pillow
streamlit
opencv-python
